Implements drag and drop
